Move out of IO namespace in main project
